requirements.txt
absl-py==0.7.1
-e git+https://github.com/SohierDane/BigQuery_Helper@8615a7f6c1663e7f2d48aa2b32c2dbcb600a440f#egg=bq_helper
cachetools==3.1.0
certifi==2019.3.9
chardet==3.0.4
Click==7.0
dash==0.41.0
dash-core-components==0.46.0
dash-html-components==0.15.0
dash-renderer==0.21.0
dash-table==3.6.0
decorator==4.3.2
Flask==1.0.2
Flask-Compress==1.4.0
google-api-core==1.9.0
google-auth==1.6.3
google-cloud-bigquery==1.11.2
google-cloud-core==0.29.1
google-resumable-media==0.3.2
googleapis-common-protos==1.5.9
gunicorn==19.9.0
idna==2.8
ipython-genutils==0.1.0
itsdangerous==1.1.0
Jinja2==2.10.1
jsonschema==2.6.0
jupyter-core==4.4.0
MarkupSafe==1.1.1
nbformat==4.4.0
numpy==1.16.2
pandas==0.24.2
plotly==3.7.1
protobuf==3.7.0
pyasn1==0.4.5
pyasn1-modules==0.2.4
pyrsistent==0.14.11
python-dateutil==2.8.0
pytz==2018.9
requests==2.21.0
retrying==1.3.3
rsa==4.0
six==1.12.0
traitlets==4.3.2
urllib3==1.24.1
Werkzeug==0.14.1
